4.GA-4/requirements.txt
# Core HTTP and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# PDF Processing
pdfminer.six>=20221105
pdfplumber>=0.10.0

# Data Processing
pandas>=2.0.0

# Web Automation
playwright>=1.42.0

# Async Support
asyncio>=3.4.3

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0 